Create EfficiencyRatioForecast
Cases are reported to not be completed in a timely manner and have been assumed to be getting worse over time. This could not be true and likely due to representatives completing their cases on time, but unable to handle their growing workloads. We are building a time series forecast using R to evaluate the efficiency of case completion in this department.
- efficiency is calculated using the formula (work output/work input)*100%
- work output/work input will translate to " cases completed/time taken to complete case"
- efficiency ratio = (cases completed/time taken to complete case) * 100%

Objective of this model is to determine the direction of efficiency as time continues on